Add --vanilla flag to all R calls as a precaution.
During development it is helpful to have an .Rprofile in the root
directory of the repository which automatically loads the package
under development when R is started in that directory. We do not want
this to happend when building the package, nor do we want any .RData
files to be loaded accidentally. Therefore it --vanilla is appended to
all calls to R during the build.