Add rotate_parameter_space().
Many of the test problems are separable in their original defintion.
It is therefore wise to rotate their parameter space.
rotate_parameter_space() is a convenince function for just this. It
randomly draws a rotation matrix R and defines a new soo_function
f_r(x) = f(Rx). See the help page for an example and further details.