Rename persp3d to plot3d.
The dependecy on 'rgl' is a pain on cluster systems. Therefore, the
persp3d method was removed and a new plot3d function with the same
signature added.

RATIONAL: Only load 'rgl' if the user asks for a 3d plot. Lose the
comfort of a method for persp3d for easier installation on headless
machines with no GL.

WHAT TO CHANGE: Replace all calls to persp3d with plot3d.